Extract generic GEMM kernel into a separate module
