+++ src/gemm/kernels/generic.rs
@@ -1,0 +1,79 @@
+use std::mem::MaybeUninit;
+use std::ops::Range;
+
+use rten_tensor::Matrix;
+
+use super::{simd_gemm, vec_count, Kernel};
+use crate::gemm::packing::{pack_a_block, pack_b_block};
+
+/// This is the base kernel that does not use architecture-specific intrinsics
+/// but is autovectorization-friendly. It is expected to perform the same as
+/// a kernel using SSE intrinsics (or equivalent).
+#[derive(Default)]
+pub struct GenericKernel {
+    _private: (),
+}
+
+impl GenericKernel {
+    const MR: usize = 8;
+
+    // The base kernel will most likely be compiled to SSE or equivalent. SSE
+    // registers are 128 bits wide = 4 x f32, so this should be a multiple of
+    // that.
+    const NR: usize = 4;
+}
+
+// Safety - Base kernel is always supported
+unsafe impl Kernel for GenericKernel {
+    fn new() -> Option<Self> {
+        Some(GenericKernel { _private: () })
+    }
+
+    fn mr(&self) -> usize {
+        Self::MR
+    }
+
+    fn nr(&self) -> usize {
+        Self::NR
+    }
+
+    fn name(&self) -> &'static str {
+        "base"
+    }
+
+    fn pack_a_block(
+        &self,
+        out: &mut [MaybeUninit<f32>],
+        a: Matrix,
+        rows: Range<usize>,
+        cols: Range<usize>,
+    ) {
+        pack_a_block::<{ Self::MR }>(out, a, rows, cols);
+    }
+
+    fn pack_b_block(
+        &self,
+        out: &mut [MaybeUninit<f32>],
+        b: Matrix,
+        rows: Range<usize>,
+        cols: Range<usize>,
+    ) {
+        pack_b_block::<{ Self::NR }>(out, b, rows, cols);
+    }
+
+    unsafe fn kernel(
+        &self,
+        tile_ptr: *mut f32,
+        tile_row_stride: usize,
+        a: &[f32],
+        b: &[f32],
+        depth: usize,
+        alpha: f32,
+        beta: f32,
+    ) {
+        const MR: usize = GenericKernel::MR;
+        const NR: usize = GenericKernel::NR;
+        const NR_REGS: usize = vec_count::<f32>(NR);
+        simd_gemm::<f32, MR, NR_REGS>(tile_ptr, tile_row_stride, a, b, depth, alpha, beta);
+    }
+}
